refactoring the structure of the components so they are not all at the top level, they will be nested according to where they lie in the web app. for example the header is in app-->dashboard-->header
