+++ CityScopeAI/src/app/dashboard/dashboard.component.css
