+++ CityScopeAI/src/app/dashboard/dashboard.component.html
