+++ CityScopeAI/src/app/dashboard/dashboard.component.ts
