+++ CityScopeAI/src/app/dashboard/data-tables/chart-handler.ts
