+++ CityScopeAI/src/app/dashboard/data-tables/data-tables.component.html
