+++ CityScopeAI/src/app/dashboard/header/header.component.css
