+++ CityScopeAI/src/app/dashboard/header/header.component.spec.ts
