+++ CityScopeAI/src/app/dashboard/heatmap/heatmap-page.component.spec.ts
