+++ CityScopeAI/src/app/dashboard/heatmap/heatmap-page.component.ts
