+++ CityScopeAI/src/app/dashboard/heatmap/heatmap/heatmap.component.spec.ts
