+++ CityScopeAI/src/app/dashboard/heatmap/heatmap/heatmap.component.ts
