+++ CityScopeAI/src/app/dashboard/heatmap/state-map/state-map.component.ts
