+++ CityScopeAI/src/app/dashboard/heatmap/zipcode-map/zipcode-map.component.html
