+++ CityScopeAI/src/services/firebase.service.spec.ts
